collection proxy models property mixin
