+++ addon/properties/relations/proxies/has-many-collection.js
@@ -2,29 +2,13 @@ import Ember from 'ember';
 import ValueWillDestroyMixin from './util/value-will-destroy-mixin';
 import { LoadableRelationLoaderStateMixin } from './util/relation-loader-state-mixin';
 import ContextPropertiesMixin from './util/context-properties-mixin';
+import CollectionModelsMixin from './util/collection-models-mixin';
 
 const {
-  computed,
   computed: { reads },
-  A
 } = Ember;
 
-const models = () => {
-  return computed('_relation.internalModels.[]', function() {
-    let relation = this._relation;
-    let modelClass = relation.relationshipModelClass;
-    let allModels = A(relation.internalModels);
-    let filtered;
-    if(modelClass) {
-      filtered = A(allModels.filter(internal => internal.definition.is(modelClass)));
-    } else {
-      filtered = allModels;
-    }
-    return A(filtered.map(internal => internal.getModel()));
-  }).readOnly();
-};
-
-const matches = () => {
+export const matches = () => {
   return reads('models');
 };
 
@@ -32,14 +16,17 @@ const content = () => {
   return reads('matches').readOnly();
 };
 
+
 export default Ember.ArrayProxy.extend(
   LoadableRelationLoaderStateMixin,
   ValueWillDestroyMixin,
-  ContextPropertiesMixin, {
+  ContextPropertiesMixin,
+  CollectionModelsMixin, {
 
   _relation: null,
 
-  models: models(),
+  query: null,
+
   matches: matches(),
   content: content(),
